feat: upgrade graphql-upload to 15.0.1
They changed the export to be ESM, so we must use the whole name with
extensions, changing the `module` in tsconfig.json and the import
lines must include the whole path with extension, just need to disable
eslint import/extensions for such lines.

The typescript is declared as JSDoc, thus tsconfig.json was also
changed to include `maxNodeModuleJsDepth: 10`, otherwise it wasn't
being found.

Convert graphql-upload to a direct runtime dependency so we're fixed
on the correct version and our users don't have to bother to align
that.